ONLINE

+++ src/main/java/co/edu/unbosque/entity/Categoria.java
@@ -16,7 +16,7 @@ public class Categoria implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name="id", nullable=false)
-	private short id;
+	private Long id;
 
 	@Column(name="estado")
 	private byte estado;
@@ -27,11 +27,11 @@ public class Categoria implements Serializable {
 	public Categoria() {
 	}
 
-	public short getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(short id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
